Bug fixes
when going into level 1 pressing k was not working sometimes. Added a bool for if your inside level area. press space bar to go in. Other bugs included spawning in the air (now fixed). changed the create player script to not look for a haskey the haskey was causing the player to spawn in the air.