Hover over item works!
added code to make the items mesh change to the mesh of the item, changes the display name to the name of the item, changes the material to a highlighted one and back to the original material. Functionality of each powerup can be added now. Make sure you save the data and test in game. Skin model changes are possible as well.