fixed enemy projectile
Enemy projectile if hits player will destroy and remove from list to spawn another. however if the projectile gets destroyed naturally after 3 seconds it does not get removed from the list. This needs to be fixed so if it is destroyed it will remove from the list for the enemy to attack again.