Boss Projectile Functions
projectile fires, kills player, destroys itself if not colliding with other, destroys itself if collided with ground or player. Creates spawn factory if not one, if one is out it wont make another until its destroyed but will still attack the player.

To Do:
particle effects,
spawn the enemies when factory is out.
particles for enemies,
damage taken and received from enemy to player,
Projectile damage when hitting player.