enemy factory rotation fix
fixed an issue where the enemies spawned would be rotated on the angle of projectiles shot. this is fixed now. also added small fixes to demo world 2, damage now works in there and world 5 onward needs a load in scrip, this script should be changed to a loading screen once all the levels are made and working in order as they should.
you got this jesse its ok!!